Attempt lab
Tried to do the following user stories:
As a customer
I want to be told whether my item has been dispensed
So that I can know whether my transaction has completed

Could not finish it.
I had zero understanding of the first day's material and the morning of
the second day which made it difficult to achieve this goal.
I though it was taught well but something must have just not clicked in
the very beginning and I got stuck straight away.
What helped me was seeing the pattern with the previous given components

Thank you for teaching us!
Wish I could have impressed you to work with you guys!